Misc: Switch to /W4 with MSVC, and fix all warnings

+++ pcsx2/CDVD/InputIsoFile.cpp
@@ -183,8 +183,8 @@ void InputIsoFile::_init()
 	m_read_inprogress = false;
 	m_read_count = 0;
 	ReadUnit = 0;
-	m_current_lsn = -1;
-	m_read_lsn = -1;
+	m_current_lsn = static_cast<u32>(-1);
+	m_read_lsn = static_cast<uint>(-1);
 	m_reader = NULL;
 }
